requirements.txt
# Core dependencies for Fiber Bundle Analysis
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
transformers>=4.21.0
statsmodels>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.64.0

# Configuration and data handling
pyyaml>=6.0
pathlib2>=2.3.0; python_version < "3.4"

# Extended dependencies for advanced features
# ==========================================

# Modern LLM support
sentence-transformers>=2.2.0    # Sentence transformer models
accelerate>=0.20.0              # Optimized model loading

# Large-scale dataset processing
datasets>=2.10.0                # HuggingFace datasets
wikipedia-api>=0.5.4            # Wikipedia data access
nltk>=3.8                       # Natural language processing
beautifulsoup4>=4.11.0          # Web scraping
requests>=2.28.0                # HTTP requests

# Scalable and distributed computing
joblib>=1.2.0                   # Parallel processing
psutil>=5.9.0                   # System monitoring

# Advanced visualization
plotly>=5.11.0                  # Interactive visualizations
umap-learn>=0.5.0               # UMAP dimensionality reduction

# Data storage and manipulation
h5py>=3.7.0                     # Efficient data storage

# Optional API dependencies
# =========================
# openai>=1.0.0                 # For OpenAI models
# anthropic>=0.3.0              # For Claude models

# Optional advanced features
# ==========================
# dask[complete]>=2023.3.0      # Distributed computing
# ray>=2.3.0                    # Distributed AI/ML workloads
# numba>=0.56.0                 # JIT compilation
# faiss-cpu>=1.7.3              # Fast similarity search

# Development dependencies (optional)
# ===================================
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0
# jupyter>=1.0.0
# ipykernel>=6.0.0